The width argument for %.*d must have type int, not size_t. This bug was introduced in r19014.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19074 06827809-a52a-0410-b366-d66718629ded

+++ apps/snmpnetstat/inet6.c
@@ -413,7 +413,7 @@ inet6print(unsigned char *in6, int port, const char *proto, int local)
 
 	struct servent *sp = NULL;
 	char line[80], *cp;
-	size_t width;
+	unsigned width;
 	int len = sizeof line;
 
 	width = Aflag ? 12 : 16;
